Move products page related code to the 'products' package
